% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics}
\alias{analytics}
\alias{analytics.viz}
\alias{analytics.analytics}
\alias{analytics.scroll}
\alias{analytics.click}
\title{Generic function that will apply a particular analytics
approach to a vizlab object}
\usage{
analytics(viz)

\method{analytics}{viz}(viz)

\method{analytics}{analytics}(viz)

\method{analytics}{scroll}(viz)

\method{analytics}{click}(viz)
}
\arguments{
\item{viz}{vizlab object or identifier}
}
\value{
vizlab object decorated with analytics
}
\description{
Generic function that will apply a particular analytics
approach to a vizlab object

Dispatch to appropriate analytics function

Catch analytics calls that have no matching type

Add a trigger for a given element to cause a analytics event
when it is scrolled to

Add an event handler for a user clicking on a visualization
that causes an analytics event to occur
}
