% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{as.analytics}
\alias{as.analytics}
\title{Coerce vizlab object to include analytics}
\usage{
as.analytics(viz, ...)
}
\arguments{
\item{viz}{vizlab object}

\item{...}{unused, left for possible pass-through}
}
\description{
Coerce vizlab object to include analytics
}
