% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{as.fetcher}
\alias{as.fetcher}
\title{Coerce vizlab object to a fetcher type}
\usage{
as.fetcher(viz, ...)
}
\arguments{
\item{viz}{vizlab object}

\item{...}{not used, following convention}
}
\value{
vizlab fetcher object
}
\description{
Coerce vizlab object to a fetcher type
}

