% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{as.global}
\alias{as.global}
\title{Coerce vizlab object to a global type}
\usage{
as.global(viz, ...)
}
\arguments{
\item{viz}{vizlab object}

\item{...}{not used, following convention}
}
\value{
vizlab global object
}
\description{
Coerce vizlab object to a global type
}
