% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{as.parameter}
\alias{as.parameter}
\title{Coerce vizlab object to a parameter type}
\usage{
as.parameter(viz, ...)
}
\arguments{
\item{viz}{vizlab object}

\item{...}{not used, following convention}
}
\value{
vizlab parameter object
}
\description{
Coerce vizlab object to a parameter type
}
