% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{as.publisher}
\alias{as.publisher}
\title{coerce to a publisher}
\usage{
as.publisher(viz, ...)
}
\arguments{
\item{viz}{object describing publisher}

\item{...}{not used, just for consistency}
}
\description{
coerce to a publisher
}

