% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{as.reader}
\alias{as.reader}
\title{Treat viz object as a reader
Doesn't need to be generic at this point (ever?)}
\usage{
as.reader(viz, ...)
}
\arguments{
\item{viz}{object to coerce to reader}

\item{...}{does nothing, following pattern}
}
\description{
Treat viz object as a reader
Doesn't need to be generic at this point (ever?)
}

