% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{as.template}
\alias{as.template}
\title{Coerce to template type publisher}
\usage{
as.template(viz, ...)
}
\arguments{
\item{viz}{vizlab object to coerce}
}
\description{
Coerce to template type publisher
}

