% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{as.visualizer}
\alias{as.visualizer}
\title{Coerce to a visualizer}
\usage{
as.visualizer(viz, ...)
}
\arguments{
\item{viz}{vizlab object}

\item{...}{not used, follows convention}
}
\value{
visualizer object
}
\description{
Coerce to a visualizer
}
