% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizlab.R
\name{as.viz}
\alias{as.viz}
\alias{as.viz.character}
\alias{as.viz.list}
\title{Coerce viz objects}
\usage{
as.viz(x, ...)

\method{as.viz}{character}(x, ...)

\method{as.viz}{list}(x, ...)
}
\arguments{
\item{x}{object to coerce to viz}

\item{...}{unused but following as. pattern}

\item{x}{character vector to convert to vizlab}

\item{...}{unused}
}
\description{
Coerce viz objects

Convert character (id) to vizlab object

Convert list into vizlab object
}

