% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{buildContext}
\alias{buildContext}
\title{Build context for templating}
\usage{
buildContext(context, dependencies)
}
\arguments{
\item{context}{context list}

\item{dependencies}{list of dependency ids}
}
\value{
list of context with dependencies injected
}
\description{
Build context for templating
}
