% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{checkAndInstallNodeModule}
\alias{checkAndInstallNodeModule}
\title{Check if a node package is installed already or not. If not, install it}
\usage{
checkAndInstallNodeModule(module, addSave = TRUE)
}
\arguments{
\item{module}{character string of the node module to install,
e.g. `webpack`, `d3`, etc}

\item{addSave}{logical indicating whether `--save` should be inlcuded 
in the install command}
}
\description{
Check if a node package is installed already or not. If not, install it
}
