% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{checkInstalled}
\alias{checkInstalled}
\title{Check that a system library is installed}
\usage{
checkInstalled(name)
}
\arguments{
\item{name}{string of the system library to look for}
}
\description{
Check that a system library is installed
}
