% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{checkRelVizUrl}
\alias{checkRelVizUrl}
\title{function to keep the absolute viz path or make the path relative}
\usage{
checkRelVizUrl(path)
}
\arguments{
\item{path}{the path defined in the viz object}
}
\description{
function to keep the absolute viz path or make the path relative
}
