% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkRequired}
\alias{checkRequired}
\title{Check required fields}
\usage{
checkRequired(x, required)
}
\arguments{
\item{x}{object to check for fields}

\item{required}{list of fields required}
}
\description{
Check required fields
}
