% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{checkThumbCompliance}
\alias{checkThumbCompliance}
\title{helper to check thumbnail compliance}
\usage{
checkThumbCompliance(width, height, size, thumbType)
}
\arguments{
\item{width}{numeric pixal width of image}

\item{height}{numeric pixal height of image}

\item{size}{numeric file size in bytes}

\item{thumbType}{char Type of thumbnail, could be "facebook", "twitter", "landing", "main"}
}
\description{
helper to check thumbnail compliance
}
