% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackages.R
\name{checkVizPackages}
\alias{checkVizPackages}
\title{Report on the installation status of packages needed by the vizlab project}
\usage{
checkVizPackages(newer = warning, older = stop, absent = stop)
}
\arguments{
\item{newer}{the function to call on a message about too-new packages (if 
there are any)}

\item{older}{the function to call on a message about too-old packages (if 
there are any)}

\item{absent}{the function to call on a message about missing packages (if 
there are any)}
}
\description{
Throw messages, warnings, errors, etc. if any packages are missing or newer 
or older than required/suggested in the viz.yaml
}
