% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake.R
\name{collectItemArguments}
\alias{collectItemArguments}
\title{Read all arguments declared for all items into an R list}
\usage{
collectItemArguments(viz.yaml)
}
\arguments{
\item{viz.yaml}{the name of the viz.yaml. should always be 'viz.yaml'.
Included as an argument so remake can make the dependency on the viz.yaml
explicit.}
}
\description{
This function is used to create dependencies for individual remake items so
that those items are rebuilt when their viz arguments change. The returned
list excludes resources, which aren't listed explicitly in the viz.yaml and
therefore can't have arguments.
}
