% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{convb}
\alias{convb}
\title{Took code from here:
https://stackoverflow.com/questions/10910688/converting-kilobytes-megabytes-etc-to-bytes-in-r}
\usage{
convb(x)
}
\arguments{
\item{x}{character representing Xbytes to convert to bytes}
}
\description{
Took code from here:
https://stackoverflow.com/questions/10910688/converting-kilobytes-megabytes-etc-to-bytes-in-r
}
