% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createBlockMakefile}
\alias{createBlockMakefile}
\title{Create a single makefile from the information in viz.yaml}
\usage{
createBlockMakefile(block = c("fetch", "process", "visualize", "publish"),
  outfile)
}
\arguments{
\item{block}{character name of the block in the viz.yaml for which to create a
makefile}

\item{outfile}{filename where the makefile should be written}
}
\description{
Uses information in the corresponding block of viz.yaml to create the
makefiles
}

