% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createCleanRule}
\alias{createCleanRule}
\title{Create the rules to clean out the make-relevant parts of the project}
\usage{
createCleanRule()
}
\description{
All that's deleted should be re-makeable
}
\keyword{internal}

