% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createConfigCall}
\alias{createConfigCall}
\title{Create the rules for the content make file}
\usage{
createConfigCall()
}
\description{
Content make file to re-build yaml files
}
\keyword{internal}
