% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createContentMakefile}
\alias{createContentMakefile}
\title{Create the content make file}
\usage{
createContentMakefile()
}
\description{
Content make file to re-build yaml files
}
\keyword{internal}
