% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeConfigRules}
\alias{createMakeConfigRules}
\title{Create the rules for the content make file}
\usage{
createMakeConfigRules()
}
\description{
Content make file to re-build yaml files
}
\keyword{internal}
