% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeDirs}
\alias{createMakeDirs}
\title{Create directories mentioned in the makefile}
\usage{
createMakeDirs(makefile)
}
\arguments{
\item{makefile}{character string containing the makefile}
}
\description{
Seeks specific directories that are expected to appear in some or all
visualization projects. Creates those for which the makefile indicates a need
}
