% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeExprRule}
\alias{createMakeExprRule}
\title{Create text for a make rule that calls an R expression}
\usage{
createMakeExprRule(target, depends = c(), expr)
}
\arguments{
\item{target}{character: the target name}

\item{depends}{character vector: the dependencies}

\item{expr}{character vector of one or more R commands}
}
\description{
The R expression will be called by R.exe -e
}
\keyword{internal}
