% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakefiles}
\alias{createMakefiles}
\title{Create fetch.make, etc. from the information in viz.yaml}
\usage{
createMakefiles(blocks = c("fetch", "process", "visualize", "publish"),
  profilepath = ".", display = FALSE)
}
\arguments{
\item{blocks}{character vector of names of blocks in the viz.yaml for which
to make makefiles}

\item{profilepath}{filepath for profile.yaml}

\item{display}{logical. [also] print the makefiles to the console?}
}
\description{
Uses information in the corresponding block of viz.yaml to create the
makefiles
}

