% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeMacros}
\alias{createMakeMacros}
\title{Create the macros section}
\usage{
createMakeMacros(profilepath)
}
\arguments{
\item{profilepath}{filepath for profile.yaml}
}
\description{
Create a character string defining the macros to include in every makefile
}

