% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeRulePair}
\alias{createMakeRulePair}
\title{Make the common rule pair relating IDs to files to commands}
\usage{
createMakeRulePair(item.info, block, concat = TRUE)
}
\arguments{
\item{item.info}{viz.yaml item info as from \code{getContentInfo}}

\item{block}{length 1 character designating the block for which to make the
rule pair}

\item{concat}{logical. Should the rules be concatenated into a single string
(TRUE) or left as a list of two rules (FALSE)?}
}
\description{
Make the common rule pair where the item ID is a symbolic target that depends
on the file, and the file depends on the scripts, args, dependencies, etc.
listed in item.info
}

