% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeRunmakeRules}
\alias{createMakeRunmakeRules}
\title{Create the rules to run/make/update the non-top makefiles}
\usage{
createMakeRunmakeRules()
}
\description{
Each makefile gets a rule to run it and a rule to create the makefile
}
\keyword{internal}
