% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeShellRule}
\alias{createMakeShellRule}
\title{Create text for a make rule that calls a shell command}
\usage{
createMakeShellRule(target, depends = c(), cmd)
}
\arguments{
\item{target}{character: the target name}

\item{depends}{character vector: the dependencies}

\item{cmd}{character vector of one or more shell commands}
}
\description{
The shell command will be called exactly as-is
}
\keyword{internal}

