% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeVizRepo.R
\name{createNewLabels}
\alias{createNewLabels}
\title{Create new labels on a repository from a JSON file.}
\usage{
createNewLabels(label_json, repo_name, org = "USGS-VIZLAB",
  ctx = get.github.context())
}
\arguments{
\item{label_json}{file path indicating the JSON file to be used to define what labels to create}

\item{repo_name}{string, name for the new repository}

\item{org}{string, GitHub organization to create repository. Defaults to "USGS-VIZLAB"}

\item{ctx}{GitHub context for authentication, see \link[grithub]{get.github.context}}
}
\description{
Create new labels on a repository from a JSON file.
}
