% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeVizRepo.R
\name{createNewRepo}
\alias{createNewRepo}
\title{Create a new repository for a specified organization.}
\usage{
createNewRepo(repo_name, description, org = "USGS-VIZLAB",
  ctx = get.github.context())
}
\arguments{
\item{repo_name}{string, name for the new repository}

\item{description}{string, short description of the repository}

\item{org}{string, GitHub organization to create repository. Defaults to "USGS-VIZLAB"}

\item{ctx}{GitHub context for authentication, see \link[grithub]{get.github.context}}
}
\description{
Create a new repository for a specified organization.
}
