% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.yaml.R
\name{createProfile.Linux}
\alias{createProfile.Linux}
\title{Create user profile yaml file for Linux operating systems}
\usage{
createProfile.Linux(file_dir)
}
\arguments{
\item{file_dir}{existing file directory for where to save the profile.yaml file}
}
\description{
Create user profile yaml file for Linux operating systems
}
\keyword{internal}

