% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.yaml.R
\name{createProfile.Mac}
\alias{createProfile.Mac}
\title{Create user profile yaml file for Mac operating systems}
\usage{
createProfile.Mac(directory)
}
\arguments{
\item{directory}{existing file directory for where to save the profile.yaml file}
}
\description{
Create user profile yaml file for Mac operating systems
}
\keyword{internal}

