% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.yaml.R
\name{createProfile.Windows}
\alias{createProfile.Windows}
\title{Create user profile yaml file for Windows operating systems}
\usage{
createProfile.Windows(directory)
}
\arguments{
\item{directory}{existing file directory for where to save the profile.yaml file}
}
\description{
Create user profile yaml file for Windows operating systems
}
\keyword{internal}

