% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createShortcuts}
\alias{createShortcuts}
\title{Create the top shortcuts}
\usage{
createShortcuts()
}
\description{
This includes an internal step "content" which is needed for triggering
a rebuild if the viz.yaml file is changed.
}
\keyword{internal}
