% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createTopMakefile}
\alias{createTopMakefile}
\title{Create the top-level makefile, which runs/makes/updates the other makefiles}
\usage{
createTopMakefile(outfile = "Makefile")
}
\arguments{
\item{outfile}{filename where the makefile should be written}
}
\description{
Declares the dependency of the other makefiles on the viz.yaml
}

