% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{doExport}
\alias{doExport}
\title{Return logical declaring whether file export (copying to target) should be
performed}
\usage{
doExport(viz, default = TRUE)
}
\arguments{
\item{viz}{vizlab object}

\item{default}{logical for default export if not specified}
}
\value{
logical should this object be exported
}
\description{
Return logical declaring whether file export (copying to target) should be
performed
}
