% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreferences.R
\name{exceededTimeToLive}
\alias{exceededTimeToLive}
\title{Return logical: has the time to live been exceeded for this viz item}
\usage{
exceededTimeToLive(id)
}
\arguments{
\item{id}{the character id of a viz item in the "fetch" block}
}
\description{
Return logical: has the time to live been exceeded for this viz item
}
