% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{expandDependencies}
\alias{expandDependencies}
\title{Private function to expand dependencies by appropriately publishing
or reading}
\usage{
expandDependencies(x)
}
\arguments{
\item{x}{item to expand}
}
\description{
Private function to expand dependencies by appropriately publishing
or reading
}
