% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{fetchTimestamp.character}
\alias{fetchTimestamp.character}
\title{Routes the timestamp fetching to a more specific fetcher.}
\usage{
\method{fetchTimestamp}{character}(viz)
}
\arguments{
\item{viz}{the identifier for a fetch item in viz.yaml}
}
\description{
This particular method is not really a timestamp fetcher, just an
intermediary, and so is not documented in [fetchTimestamp()].
}
