% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{fetchTimestamp.url}
\alias{fetchTimestamp.url}
\title{check a URL for timestamp}
\usage{
\method{fetchTimestamp}{url}(viz)
}
\description{
@rdname fetchTimestamp
 @importFrom httr HEAD
 @importFrom httr headers
 @export
}

