% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfileInfo.R
\name{findProfileYaml}
\alias{findProfileYaml}
\title{Find where the profile.yaml file is saved.}
\usage{
findProfileYaml()
}
\description{
Searches two directories, '~/.vizlab/' and './vizlab/', for a file called 
'profile.yaml'. If found, returns the file path where it was found; otherwise
throws an error.
}
