% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gatherDependencyList}
\alias{gatherDependencyList}
\title{Pull together vizlab object dependencies}
\usage{
gatherDependencyList(...)
}
\arguments{
\item{...}{dependencies to gather}
}
\description{
Pull together vizlab object dependencies
}
