% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{generatePackageJsonFile}
\alias{generatePackageJsonFile}
\title{Create the package.json file}
\usage{
generatePackageJsonFile(file, webpack.config)
}
\arguments{
\item{file}{the file name to save the output with .json ending}

\item{webpack.config}{list generated by `getWebpackConfig()`

Could switch out the template to a custom one.}
}
\description{
Create the package.json file
}
