% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{generateWebpackConfigFile}
\alias{generateWebpackConfigFile}
\title{Create the webpack.config.js file
Needs an input and output in webpack.config}
\usage{
generateWebpackConfigFile(file, webpack.config)
}
\arguments{
\item{file}{the file name to save the output with .js ending}

\item{webpack.config}{list generated by `getWebpackConfig()`

Could switch out the template to a custom one.}
}
\description{
Create the webpack.config.js file
Needs an input and output in webpack.config
}
