% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAutoargs.R
\name{getAutoargs}
\alias{getAutoargs}
\alias{getAutoargs.character}
\alias{getAutoargs.list}
\title{Prepare the arguments to pass to a fetch, read, process, or
visualize}
\usage{
getAutoargs(data.info, fun = c("write", "read"), verbose = TRUE)

\method{getAutoargs}{character}(data.info, ...)

\method{getAutoargs}{list}(data.info, fun = c("write", "read"),
  verbose = TRUE)
}
\arguments{
\item{data.info}{character viz.id or list as in getContentInfo(vizid)}

\item{fun}{data mode: read or write? write is the default}

\item{verbose}{logical. should there be a message that we're using autoargs?}

\item{...}{args passed on to other getAutoargs methods}
}
\description{
Prepare the arguments to pass to a fetch, read, process, or
visualize
}

