% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentInfo.R
\name{getBlocks}
\alias{getBlocks}
\title{Read vizlab config document and return specific blocks}
\usage{
getBlocks(block = c("vizlab", "info", "fetch", "process", "visualize",
  "publish"))
}
\arguments{
\item{block}{character vector of blocks to read}
}
\description{
Read vizlab config document and return specific blocks
}

