% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentInfo.R
\name{getContentInfo}
\alias{getContentInfo}
\title{Get information on a content item from the viz.yaml}
\usage{
getContentInfo(viz.id, block = c("parameter", "fetch", "process",
  "visualize", "publish", "resource"), no.match = c("stop", "NA"))
}
\arguments{
\item{viz.id}{a length-1 character string giving the ID for a single content
item to return}

\item{block}{the name or names of content blocks whose items should be
included in the query}

\item{no.match}{what to do if the viz.id is not found: either 'stop' (throw
error) or 'NA' (return NA)}
}
\description{
Returns a single content item. Reads from the viz.yaml in the current working
directory
}
\seealso{
getContentInfos
}
