% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentInfo.R
\name{getContentInfos}
\alias{getContentInfos}
\title{Get information on one or more content items from the viz.yaml}
\usage{
getContentInfos(viz.id, block = c("fetch", "process", "visualize", "publish",
  "resource"))
}
\arguments{
\item{viz.id}{character[s] giving specific IDs of items to return, or missing
to return all content items}

\item{block}{the name or names of content blocks whose items should be
included in the query}
}
\description{
Returns a list of content items, each of which contains a list of information
key-value pairs. Reads from the viz.yaml in the current working directory
}
\seealso{
getContentInfo
}
