% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLog.R
\name{getLog}
\alias{getLog}
\title{Display, optionally open, and return the contents of a log file}
\usage{
getLog(id, open = FALSE, tail = TRUE)
}
\arguments{
\item{id}{character the viz id, eventually any possible target listed in the
build output.}

\item{open}{logical. Should the file be opened in an editor?}

\item{tail}{logical. Should the tail of the file be printed to the console?}
}
\description{
Prints a message with the tail of the log file if `tail=TRUE`. Opens the file
in an editor if `open=TRUE`. Always returns the entire log file as a
character vector.
}
