% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLog.R
\name{getLogPath}
\alias{getLogPath}
\title{Get the path to the .Rout file corresponding to a viz id (or an id_timestamp)}
\usage{
getLogPath(id)
}
\arguments{
\item{id}{character the viz id, eventually any possible target listed in the
build output.}
}
\description{
Get the path to the .Rout file corresponding to a viz id (or an id_timestamp)
}
