% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPartialLibrary}
\alias{getPartialLibrary}
\title{Assemble whisker partials from vizlab package}
\usage{
getPartialLibrary()
}
\value{
list containing partials available
}
\description{
Assemble whisker partials from vizlab package
}
