% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreferences.R
\name{getPreferences}
\alias{getPreferences}
\title{Read the preferences.yaml file, if available, in the current working
directory. Return as a formatted R list.}
\usage{
getPreferences()
}
\description{
Read the preferences.yaml file, if available, in the current working
directory. Return as a formatted R list.
}
\section{Sections in preferences.yaml and the output of \code{getPreferences()}}{

\itemize{
\item \code{timetolive}: a list of difftimes, named for the corresponding viz IDs.
The format is a length-2 space-separated string describing a difftime. For
example, \code{Inf days} becomes \code{as.difftime(as.numeric("Inf"), units="days"},
\code{6 hours} becomes \code{as.difftime(as.numeric("6"), units="days")}
}
}

