% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfileInfo.R
\name{getProfileInfo}
\alias{getProfileInfo}
\title{Get information on profile settings}
\usage{
getProfileInfo(user = "local", profilepath)
}
\arguments{
\item{user}{username, probably either 'local' or 'docker', whose profile to
retrieve}

\item{profilepath}{filepath for profile.yaml}
}
\description{
Reads all values from the specified user's profile.yaml
}

