% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{getRepoNames}
\alias{getRepoNames}
\title{Get a vector a repository names for a specific organization}
\usage{
getRepoNames(org)
}
\arguments{
\item{org}{character, name of GitHub organization in which to pull out repository names}
}
\description{
Get a vector a repository names for a specific organization
}
