% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getResourceFromLibrary}
\alias{getResourceFromLibrary}
\title{Internal function to get shared resources
Implemented as a closure to avoid reloading file each time}
\usage{
getResourceFromLibrary(x, no.match = c("stop", "NA"))
}
\arguments{
\item{x}{character vector containing resource id}

\item{no.match}{what to do if the viz.id is not found: either 'stop' (throw
error) or 'NA' (return NA)}
}
\value{
vizlab object from library or \code{NULL} if it doesn't exist
}
\description{
Internal function to get shared resources
Implemented as a closure to avoid reloading file each time
}

