% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{getTemplateFromLibrary}
\alias{getTemplateFromLibrary}
\title{Internal function to get shared templates}
\usage{
getTemplateFromLibrary(x)
}
\arguments{
\item{x}{character vector template name}
}
\value{
character vector template contents
}
\description{
Internal function to get shared templates
}

