% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{getVizHTML}
\alias{getVizHTML}
\title{Populate mustache template with vizzy information}
\usage{
getVizHTML(viz_info)
}
\arguments{
\item{viz_info}{list with a list of viz information called "vizzies", e.g. list(vizzies=list(viz_info1, viz_info2))}
}
\description{
Populate mustache template with vizzy information
}

