% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{getVizInfo}
\alias{getVizInfo}
\title{Load yamls and find if the published date has passed yet}
\usage{
getVizInfo(org, repo)
}
\arguments{
\item{org}{character, name of GitHub organization in which to look for a repository}

\item{repo}{character, name of the repository to find the viz.yaml}
}
\description{
Load yamls and find if the published date has passed yet
}

