% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{getVizThumbnail}
\alias{getVizThumbnail}
\title{Get the url for the viz thumbnail}
\usage{
getVizThumbnail(org, repo)
}
\arguments{
\item{org}{character, name of GitHub organization in which to look for a repository}

\item{repo}{character, name of the repository to find the thumbnail}
}
\description{
Get the url for the viz thumbnail
}

