% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{getVizYamlUrl}
\alias{getVizYamlUrl}
\title{Get the full file path for the raw viz.yaml within a repository}
\usage{
getVizYamlUrl(org, repo)
}
\arguments{
\item{org}{character, name of GitHub organization in which to look for a repository}

\item{repo}{character, name of the repository to find the viz.yaml}
}
\description{
Get the full file path for the raw viz.yaml within a repository
}
