% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global}
\alias{global}
\alias{global.character}
\alias{global.global}
\title{Get global variables for project}
\usage{
global(viz)

\method{global}{character}(viz)

\method{global}{global}(viz)
}
\arguments{
\item{viz}{vizlab object described in viz.yaml}
}
\description{
This function should be called from the generic, \code{global()}.
}
