% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{handleMarkdown}
\alias{handleMarkdown}
\title{Replace any markdown text with rendered html}
\usage{
handleMarkdown(text)
}
\arguments{
\item{text}{character vector containing markdown text}
}
\description{
Replace any markdown text with rendered html
}

