% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeVizRepo.R
\name{initializeVizRepo}
\alias{initializeVizRepo}
\title{Initialize the GitHub repo for a new viz.}
\usage{
initializeVizRepo(repo_name, description, org = "USGS-VIZLAB",
  issue_json = file.path("inst", "issuetemplates.json"),
  label_json = file.path("inst", "labeltemplates.json"))
}
\arguments{
\item{repo_name}{string, name for the new repository}

\item{description}{string, short description of the repository}

\item{org}{string, GitHub organization to create repository. Defaults to "USGS-VIZLAB"}

\item{issue_json}{file path indicating the JSON file to be used to define what issues to create}

\item{label_json}{file path indicating the JSON file to be used to define what labels to create}
}
\value{
URL for the new GitHub repository.
}
\description{
run this command to create a GitHub repository with 
appropriate issues, description, and README.
}
