% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{locateTimestampFile}
\alias{locateTimestampFile}
\title{Where is the timestamp file}
\usage{
locateTimestampFile(id)
}
\arguments{
\item{id}{viz id needing timestamp}
}
\value{
character vector location of timestamp file
}
\description{
Where is the timestamp file
}
