% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lookupMimetype}
\alias{lookupMimetype}
\title{Use mimetype lookup to get reader}
\usage{
lookupMimetype(mimetype)
}
\arguments{
\item{mimetype}{character vector of length one with the mimetype name}
}
\value{
character vector describing the reader to be used
}
\description{
Use mimetype lookup to get reader
}
