% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeCleanRule}
\alias{makeCleanRule}
\title{Create the rules to clean out the make-relevant parts of the project}
\usage{
makeCleanRule()
}
\description{
All that's deleted should be re-makeable
}
\keyword{internal}

