% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataRemakefile.R
\name{makeDataRemakefile}
\alias{makeDataRemakefile}
\title{Create data.yaml from the information in viz.yaml}
\usage{
makeDataRemakefile()
}
\description{
Uses information in the data block of viz.yaml to create the remake file
data.yaml
}

