% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeBatchRule}
\alias{makeMakeBatchRule}
\title{Create text for a make rule using RBATCH}
\usage{
makeMakeBatchRule(target, depends = c(), fun, funargs = c(),
  scripts = c(), logfile)
}
\arguments{
\item{target}{character: the target name}

\item{depends}{character vector: the dependencies (omit callFunction.R and 
scripts; these will be automatically added)}

\item{fun}{character: the function name, to be ultimately evaluated in an 
environment where the scripts have been sourced and library(vizlab) has 
been called. if fun is missing, the command section of the make target
declaration will be omitted}

\item{scripts}{character vector: the scripts or script directories to load. 
omit 'scripts' because this will be automatically preprended}

\item{logfile}{character: the filename where the R CMD BATCH logfile should 
be saved. omit 'vizlab/make/log' because this will be automatically 
prepended}
}
\description{
Create the text for a gnu make target that runs a function using R CMD BATCH 
and callFunction.R
}
\keyword{internal}

