% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeEmptyRule}
\alias{makeMakeEmptyRule}
\title{Create text for a make rule that has no commands}
\usage{
makeMakeEmptyRule(target, depends = c())
}
\arguments{
\item{target}{character: the target name}

\item{depends}{character vector: the dependencies}
}
\description{
The resulting rule will have only a target and maybe dependencies, no
commands
}
\keyword{internal}

