% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakefiles}
\alias{makeMakefiles}
\title{Create fetch.make, etc. from the information in viz.yaml}
\usage{
makeMakefiles(blocks = c("fetch", "process", "visualize"), display = FALSE)
}
\arguments{
\item{blocks}{character vector of names of blocks in the viz.yaml for which
to make makefiles}
}
\description{
Uses information in the corresponding block of viz.yaml to create the 
makefiles
}

