% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeItem.fetch}
\alias{makeMakeItem.fetch}
\title{\code{makeMakeItem.fetch}: Make makefile rules for an item in the fetch block
of viz.yaml}
\usage{
\method{makeMakeItem}{fetch}(item.info, ...)
}
\description{
\code{makeMakeItem.fetch}: Make makefile rules for an item in the fetch block
of viz.yaml
}

