% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeItem.process}
\alias{makeMakeItem.process}
\title{\code{makeMakeItem.process}: Make makefile rules for an item in the
process block of viz.yaml}
\usage{
\method{makeMakeItem}{process}(item.info)
}
\description{
\code{makeMakeItem.process}: Make makefile rules for an item in the
process block of viz.yaml
}

