% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeRunmakeRules}
\alias{makeMakeRunmakeRules}
\title{Create the rules to run/make/update the non-top makefiles}
\usage{
makeMakeRunmakeRules()
}
\description{
Each makefile gets a rule to run it and a rule to make the makefile
}
\keyword{internal}

