% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeShellRule}
\alias{makeMakeShellRule}
\title{Create text for a make rule that calls a shell command}
\usage{
makeMakeShellRule(target, depends = c(), cmd)
}
\arguments{
\item{target}{character: the target name}

\item{depends}{character vector: the dependencies}

\item{cmd}{character vector of one or more shell commands}
}
\description{
The shell command will be called exactly as-is
}
\keyword{internal}

