% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{needsTimestamp}
\alias{needsTimestamp}
\title{Determine whether an item should fetch a timestamp or not}
\usage{
needsTimestamp(item.info)
}
\value{
logical: TRUE if timestamp is needed (== a
fetchTimestamp.fetcherforthisvizitem method is available), FALSE otherwise.
Also produces an error if there's a mismatch between what's declared in the
viz.yaml and whether a fetchTimestamp method appears to be available for
this fetcher.
}
\description{
Looks in the dependency scripts and on the current search path (plus vizlab)
to determine whether the fetchTimestamp method for this fetcher is actually
available somewhere. DOESN'T currently look in any packages loaded via
\code{library} within the sourced scripts (to avoid altering the current
environment without the user's permission), which means we'll get confused if
some other package implements, say, fetchTimestamp.yadayada. That's an
unlikely situation, so if we need it, we can implement it then (or the user
can call \code{library()} on that package before calling \code{createMakefiles()}).
}
\details{
Looks at the fetchTimestamp argument in this item's viz.yaml info to see what
the user expects.
}
\keyword{internal}
