% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackages.R
\name{packageStatus}
\alias{packageStatus}
\title{Identify any missing or outdated packages}
\usage{
packageStatus()
}
\description{
Compares the installed package versions to those given in the
`required-packages` section of the viz.yaml. Returns a list of two vectors,
one with the names of completely missing packages and one with the names of
out of date packages.
}
