% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{parameter}
\alias{parameter}
\alias{parameter.character}
\alias{parameter.parameter}
\title{Get parameter variables for project}
\usage{
parameter(viz)

\method{parameter}{character}(viz)

\method{parameter}{parameter}(viz)
}
\arguments{
\item{viz}{vizlab object described in viz.yaml}
}
\description{
This function should be called from the generic, \code{parameter()}.
}
