% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSources.R
\name{prepSources}
\alias{prepSources}
\title{Combines and filters one or more script files into a single script
appropriate to a single viz item}
\usage{
prepSources(..., functions, outfile)
}
\arguments{
\item{...}{length-1 character strings of R script filenames to be parsed`}

\item{functions}{length-1 character string of 1+ functions to include, or ""
to include all functions. The contents of the combined scripts can be
optionally filtered to include definitions for only those functions (or R
objects) listed here. If
there are >1 functions to include, they should be separated by commas wtihin the same
string, e.g., `'fetch.cars,cars.helper'`}

\item{outfile}{full file path where the combined source file should be
written. This will usually be vizlab/remake/scripts/{vizitemname}.R}
}
\description{
Return the MD5sum of a deparsed script - this shouldn't change if whitespace
or comments change but should change if there's anything substantive
}
