% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\alias{process.character}
\alias{process.unzip}
\title{Process raw or intermediate data products into intermediate or final data
products}
\usage{
process(viz)

\method{process}{character}(viz)

\method{process}{unzip}(viz)
}
\arguments{
\item{viz}{vizlab object described in viz.yaml}
}
\description{
This function should be called from the generic, \code{process()}.
Process raw or intermediate data products into intermediate or final data
products

\code{process.unzip} unzip a zip file
}

