% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{publishLandingPage}
\alias{publishLandingPage}
\title{Get information about other visualizations from VIZLAB to generate the landing page}
\usage{
publishLandingPage()
}
\description{
the github package from the GitHub repo rgithub (https://github.com/cscheid/rgithub)
was moved to GRAN and called grithub
}
