% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\alias{readData.character}
\alias{readData.viz}
\alias{readData.tabular}
\alias{readData.yaml}
\alias{readData.excel}
\alias{readData.rds}
\alias{readData.filepath}
\alias{readData.parameter}
\alias{readData.folder}
\alias{readData.txt}
\alias{readData.md}
\alias{readData.inline}
\alias{readData.json}
\alias{readData.shp}
\alias{readData.svg}
\title{Ensure data is on the local file system}
\usage{
readData(viz)

\method{readData}{character}(viz)

\method{readData}{viz}(viz)

\method{readData}{tabular}(viz)

\method{readData}{yaml}(viz)

\method{readData}{excel}(viz)

\method{readData}{rds}(viz)

\method{readData}{filepath}(viz)

\method{readData}{parameter}(viz)

\method{readData}{folder}(viz)

\method{readData}{txt}(viz)

\method{readData}{md}(viz)

\method{readData}{inline}(viz)

\method{readData}{json}(viz)

\method{readData}{shp}(viz)

\method{readData}{svg}(viz)
}
\arguments{
\item{viz}{vizlab object or identifier}
}
\value{
single value or list of values
}
\description{
This function should be called from the generic, \code{readData()}. Reads
data from a file into R format. If you just want the filepath so you can do
it yourself, specify "customReader: filepath" in the viz.yaml

Can either be a single markdown file, specified with mimetype: text/markdown
or markdown contained in yaml with mimetype: text/yaml

assumes only one .shp file in zip folder
}
\details{
If simple markdown is used, return will be the rendered html, otherwise it
will be a list of html fragments named with the keys from the yaml.
}
