% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDepends.R
\name{readDepends}
\alias{readDepends}
\alias{readDepends.character}
\alias{readDepends.list}
\alias{readDepends.viz}
\title{read multiple dependency datasets into environment}
\usage{
readDepends(viz)

\method{readDepends}{character}(viz)

\method{readDepends}{list}(viz)

\method{readDepends}{viz}(viz)
}
\arguments{
\item{viz}{vizlab object, list, or vizlab identifier}
}
\value{
a list of data objects that are named according to depends
}
\description{
This function should be called from the generic, \code{readDepends()}. Reads
all dependency data from files into R format.
}
\examples{
wd <- getwd()
setwd(system.file(package = 'vizlab','testviz'))
#Read dependencies from list or viz object:
viz.data <- readDepends(list(depends = 'MayflyNymph'))
viz.data[["MayflyNymph"]]

setwd(wd)
}
\seealso{
readData
}
