% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{readOldTimestamp}
\alias{readOldTimestamp}
\title{Read an old timestamp for viz}
\usage{
readOldTimestamp(viz)
}
\arguments{
\item{viz}{vizlab object}
}
\value{
POSIXct timestamp for vizlab object
}
\description{
Read an old timestamp for viz
}

