% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplate}
\alias{readTemplate}
\title{Reads a template file and returns the contents}
\usage{
readTemplate(x)
}
\arguments{
\item{x}{character vector representing template by name or file path}
}
\value{
character vector containing template contents
}
\description{
Reads a template file and returns the contents
}
