% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{relativePath}
\alias{relativePath}
\title{Get webapp path}
\usage{
relativePath(file)
}
\arguments{
\item{file}{path to file being exported}
}
\value{
character vector describing relative path
}
\description{
Get webapp path
}
