% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublish.R
\name{remove_publish_target}
\alias{remove_publish_target}
\title{Helper to unpublish. similar to remake:::remake_remove_target, but without file deletion}
\usage{
remove_publish_target(obj, target_name)
}
\arguments{
\item{obj}{argument as in remake:::remake_remove_target}

\item{target_name}{argument as in remake:::remake_remove_target}
}
\description{
Helper to unpublish. similar to remake:::remake_remove_target, but without file deletion
}
\keyword{internal}
