% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\alias{render.template}
\title{Render contents to file}
\usage{
render(viz, data, file)

\method{render}{template}(viz, data, file = NULL)
}
\arguments{
\item{viz}{vizlab object}

\item{data}{to render}

\item{file}{file to render to}
}
\description{
Render contents to file

Renter template, return output if file not specified
}
