% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replaceOrAppend}
\alias{replaceOrAppend}
\title{Append second list to first with overwrites}
\usage{
replaceOrAppend(x, y)
}
\arguments{
\item{x}{list template to be filled by values of another list}

\item{y}{list to overwrite missing values or append to first list}
}
\value{
list containing merged values from both x and y
}
\description{
Append second list to first with overwrites
}
