% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authRemote.R
\name{sbAuthenticate}
\alias{sbAuthenticate}
\title{Login and if desired, store, credentials for sciencebase}
\usage{
sbAuthenticate()
}
\description{
Login and if desired, store, credentials for sciencebase
}

