% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setupFoldersForFile}
\alias{setupFoldersForFile}
\title{Sets up folders so file can be written without warnings}
\usage{
setupFoldersForFile(file)
}
\arguments{
\item{file}{file that is about to be written}
}
\description{
Sets up folders so file can be written without warnings
}
