% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceScripts.R
\name{sourceScripts}
\alias{sourceScripts}
\title{Source R scripts}
\usage{
sourceScripts(scripts, verbose = TRUE)
}
\arguments{
\item{scripts}{character or character vectory. If an element of scripts is a
directory, sources the contents of that directory. Otherwise sources the
script.}

\item{verbose}{logical. Declare which scripts are being sourced via messages?}
}
\description{
Sources the script or scripts indicated.
}

