% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\alias{template.character}
\title{Reads a template file and returns the contents}
\usage{
template(viz)

\method{template}{character}(viz)
}
\arguments{
\item{viz}{character vector representing template by name or file path}
}
\value{
character vector containing template contents
}
\description{
Reads a template file and returns the contents

Read a template from file or id
}

