% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{uniqueness}
\alias{uniqueness}
\title{Grab a random number to break the cache}
\usage{
uniqueness()
}
\value{
random number between 10000 and 10000000
}
\description{
Grab a random number to break the cache
}

