% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{updateConfigInfoFile}
\alias{updateConfigInfoFile}
\title{updateConfigInfoFile}
\usage{
updateConfigInfoFile(viz.id)
}
\arguments{
\item{viz.id}{character[s] giving specific IDs of items to return, or missing
to return all content items}
}
\description{
Creates 1 file per viz id with the content info. This is necessary to 
update the project if only the viz.yaml is updated (otherwise make)
won't trigger an update.
}
