% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{updateContent}
\alias{updateContent}
\title{updateContent}
\usage{
updateContent(block = c("fetch", "process", "visualize", "publish"))
}
\arguments{
\item{block}{character name of the block in the viz.yaml for which to create a
makefile}
}
\description{
Create rds files for each id's content
}
