% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{updateThumbnails}
\alias{updateThumbnails}
\title{Get thumbnail info for sematics}
\usage{
updateThumbnails(context, thumbnails)
}
\arguments{
\item{context}{list, should be just info section}

\item{thumbnails}{list, taken from full context}
}
\description{
Get thumbnail info for sematics
}
