% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackages.R
\name{updateVizPackages}
\alias{updateVizPackages}
\title{[Re]install any missing or outdated packages}
\usage{
updateVizPackages(install.if = c("older", "missing"))
}
\arguments{
\item{install.if}{package statuses that should cause that package to be 
reinstalled. The default is recommended, but "newer" is a 
sometimes-reasonable additional option, and "perfect" is technically also 
an option (if you want to reinstall everything no matter what).}
}
\description{
For packages whose `repo` field in `required-packages` is CRAN or GRAN, the 
newly installed package will be the most current one, not the precise version
specified in the `version` field. For packages whose `repo` is github, the 
package will be installed from the specified `ref`.
}
