% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\alias{visualize.character}
\title{Visualize raw or intermediate data products into intermediate or final data
products}
\usage{
visualize(viz)

\method{visualize}{character}(viz)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{...}{other arguments passed to visualize methods. These should
include any data dependencies, named according to the viz.id of those data
items}

\item{outfile}{the filename to which the visualized data should be written}
}
\description{
This function should be called from the generic, \code{visualize()}.
Visualize raw or intermediate data products into intermediate or final data
products
}

