% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeData.R
\name{visualizeData}
\alias{visualizeData}
\alias{visualizeData.character}
\title{Visualize raw or intermediate data products into intermediate or final data 
products}
\usage{
visualizeData(viz.id, ..., outfile)

\method{visualizeData}{character}(viz.id, ..., outfile)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{...}{other arguments passed to visualizeData methods. These should 
include any data dependencies, named according to the viz.id of those data 
items}

\item{outfile}{the filename to which the visualized data should be written}
}
\description{
This function should be called from the generic, \code{visualizeData()}. 
Visualize raw or intermediate data products into intermediate or final data 
products
}

