% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.skeleton.R
\name{viz.skeleton}
\alias{viz.skeleton}
\title{vizlab viz.skelton}
\usage{
viz.skeleton(name = "cool-viz", path = ".")
}
\arguments{
\item{name}{character vector containing viz name}

\item{path}{character vector of path to create skeleton}
}
\description{
Create a skeleton for a new visualization
}
\examples{
path <- file.path(tempdir(), "test")
viz.skeleton(name="test", path=path)
file.remove(path) # cleanup

}

