% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizlab.R
\name{vizlab}
\alias{as.viz}
\alias{as.viz.character}
\alias{as.viz.list}
\alias{vizlab}
\title{Utilities for building online data visualizations}
\usage{
as.viz(x, ...)

\method{as.viz}{character}(x, ...)

\method{as.viz}{list}(x, ...)
}
\arguments{
\item{x}{object to coerce to viz.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Utilities for building online data visualizations

Coerce viz objects

Convert character (id) to vizlab object

Convert list into vizlab object
}

