% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizSkeleton.R
\name{vizSkeleton}
\alias{vizSkeleton}
\title{vizlab vizSkelton}
\usage{
vizSkeleton(name = "cool-viz", path = ".")
}
\arguments{
\item{name}{character vector containing viz name}

\item{path}{character vector of path to create skeleton}
}
\description{
Create a skeleton for a new visualization
}
\examples{
path <- file.path(tempdir(), "test")
vizSkeleton(name="test", path=path)
file.remove(path) # cleanup

}

