% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{webpacker}
\alias{webpacker}
\title{Run webpack for the project}
\usage{
webpacker(webpack_build_cfg)
}
\arguments{
\item{webpack_cfg}{character string indicating which script to run from the
webpack package.json file, e.g. "dev", "prod", "watch", etc. Must exist in
the package.json file for your viz. Defaults to \code{start}.}
}
\description{
This is called at the very end of \code{vizmake()} if you have a webpack block in your \code{viz.yaml}.
}
