% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{writeTimestamp}
\alias{writeTimestamp}
\title{Write a timestamp file with the conventions used by fetchTimestamp}
\usage{
writeTimestamp(new.timestamp, timestamp.file)
}
\description{
Should only be used outside the vizlab package if the developer is writing a 
new data fetcher
}

