% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authRemote.R
\name{authRemote}
\alias{authRemote}
\alias{authRemote.default}
\alias{authRemote.sciencebase}
\title{Log on to a remote file service}
\usage{
authRemote(fetcher, ...)

\method{authRemote}{default}(fetcher, user = "local", ...)

\method{authRemote}{sciencebase}(fetcher, user, ...)
}
\arguments{
\item{fetcher}{the name of the fetcher for which to authenticate, e.g., 
'sciencebase'}

\item{...}{other arguments passed to authRemote methods}

\item{user}{not yet well thought out, but the idea is to accept 'local' or 
'docker' and use the corresponding user profile to authenticate}
}
\description{
Uses authentication information in the ~/.vizlab driectory to log onto a remote 
file service such as ScienceBase. This function is primarily for internal use
but can be called directly to check credentials.

\code{authRemote.sciencebase} Uses the credentials stored in ~/.vizlab to log in to 
ScienceBase.  Use the storeSBcreds function to verify and store credentials.  Since vizlab
starts its own R session, an existing ScienceBase session cannot be used.
}
\details{
To add ScienceBase credentials, run the \code{storeSBcreds} function.
}

