% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{checkThumbCompliance}
\alias{checkThumbCompliance}
\title{helper to check thumbnail compliance}
\usage{
checkThumbCompliance(file, maxSize, checkHeight, checkWidth)
}
\description{
helper to check thumbnail compliance
}
