% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{checkThumbCompliance}
\alias{checkThumbCompliance}
\title{helper to check thumbnail compliance}
\usage{
checkThumbCompliance(file, thumbType)
}
\arguments{
\item{file}{char Name of thumbnail file}

\item{thumbType}{char Type of thumbnail, could be "facebook", "twitter", "landing", "main"}
}
\description{
helper to check thumbnail compliance
}
