% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMakefiles.R
\name{createMakeItem}
\alias{createMakeItem}
\alias{createMakeItem.default}
\alias{createMakeItem.fetch}
\alias{createMakeItem.makefile}
\alias{createMakeItem.process}
\alias{createMakeItem.publish}
\alias{createMakeItem.visualize}
\title{Make a collection of makefile rules appropriate to a data/figure item}
\usage{
createMakeItem(item.info, ...)

\method{createMakeItem}{default}(item.info, ...)

\method{createMakeItem}{fetch}(item.info, ...)

\method{createMakeItem}{process}(item.info, ...)

\method{createMakeItem}{visualize}(item.info, ...)

\method{createMakeItem}{publish}(item.info, ...)

\method{createMakeItem}{makefile}(item.info, ...)
}
\arguments{
\item{item.info}{viz.yaml item info as from \code{getContentInfo}}

\item{...}{other args passed to makeMakeItem methods}
}
\description{
Make a collection of makefile rules appropriate to a data/figure item

\code{createMakeItem.fetch}: create makefile rules for an item in the fetch block
of viz.yaml

\code{createMakeItem.process}: create makefile rules for an item in the
process block of viz.yaml

\code{createMakeItem.visualize}: create makefile rules for an item in the
visualize block of viz.yaml

\code{createMakeItem.publish}: create makefile rules for an item in the
publish block of viz.yaml

\code{createMakeItem.visualize}: create makefile rules for an item in the
visualize block of viz.yaml
}

