% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.yaml.R
\name{createProfile}
\alias{createProfile}
\title{Define user profile information}
\usage{
createProfile(directory = "home")
}
\arguments{
\item{directory}{location to save the profile.yaml file. There are currently 2 options: "home" or "relative"}
}
\description{
This creates a file that describes where necessary executables are located in order to run make.
This sets up a template, it is necessary to fill in the missing information
}
\examples{
\dontrun{
# Example file for Windows user[s]:
SHELL: /usr/bin/sh
R: C:/Program Files/R/R-3.3.0/bin/x64/R.exe
RSCRIPT: C:/Program Files/R/R-3.3.0/bin/x64/Rscript.exe
R_LIBS_USER: !eval >
  paste0(
    if(basename(Sys.getenv("R_USER"))=='Documents')
      Sys.getenv("R_USER")
    else
      normalizePath(file.path(Sys.getenv("R_USER"), 'Documents'), winslash='/'),
    "/R/win-library/3.3")
}
}

