% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{doExport}
\alias{doExport}
\title{Should export be performed}
\usage{
doExport(viz, default)
}
\arguments{
\item{viz}{vizlab object}

\item{default}{logical for default export if not specified}
}
\value{
logical should this object be exported
}
\description{
Should export be performed
}
