% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{export.css}
\alias{export.img}
\alias{export.js}
\alias{export.page}
\alias{export.svg}
\title{Export content to target}
\usage{
export(viz)

\method{export}{page}(viz)

\method{export}{img}(viz)

\method{export}{svg}(viz)

\method{export}{js}(viz)

\method{export}{css}(viz)
}
\arguments{
\item{viz}{object to export}
}
\description{
Determine the type and dispatch to that method to produce
files to serve up as the final viz

This returns the target location to write the page to, in
its current form this is a single page, but may result in
a list of pages named according to different formats
(tags, media type, language, etc.)
}
\details{
Currently the publish phase is responsible for exporting,
but this could be separated in the future.
}

