% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\alias{fetch.character}
\alias{fetch.file}
\alias{fetch.sciencebase}
\alias{fetch.url}
\alias{fetch.usgs_watermark}
\alias{fetch.none}
\alias{fetchTimestamp.none}
\title{Ensure data is on the local file system}
\usage{
fetch(viz)

\method{fetch}{character}(viz)

\method{fetch}{file}(viz)

\method{fetch}{sciencebase}(viz)

\method{fetch}{url}(viz)

\method{fetch}{usgs_watermark}(viz)

\method{fetch}{none}(viz)

\method{fetchTimestamp}{none}(viz)
}
\arguments{
\item{viz}{the identifier for this data item in viz.yaml}
}
\description{
This function should be called from the generic, \code{fetch()}. Fetch
data from remote locations and/or identify a copy on the local file system.
These functions assume that downloading must be attempted if the raw data
file is remote; the makefile is where caching occurs if possible.

\code{fetch.file} currently does nothing. It exists to communicate that
the .file type is an option for data items.

\code{fetch.sciencebase} downloads a file from ScienceBase.

\code{fetch.url} downloads a file from the specified URL.

\code{fetch.usgs_watermark} grabs the USGS watermark

\code{fetch.none} skips this fetch

\code{fetchTimestamp.none} skips this fetchTimestamp
}
