% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{fetchData}
\alias{fetchData}
\alias{fetchData.default}
\alias{fetchData.file}
\alias{fetchData.sciencebase}
\alias{fetchData.url}
\title{Ensure data is on the local file system}
\usage{
fetchData(viz.id, ...)

\method{fetchData}{default}(viz.id, data.info, ...)

\method{fetchData}{file}(viz.id, data.info, ...)

\method{fetchData}{sciencebase}(viz.id, data.info, ...)

\method{fetchData}{url}(viz.id, data.info, ...)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{...}{other arguments passed to fetchData methods}

\item{data.info}{content information for this viz.id from the viz.yaml}
}
\description{
This function should be called from the generic, \code{fetchData()}. Fetch
data from remote locations and/or identify a copy on the local file system.
These functions assume that downloading must be attempted if the raw data
file is remote; the makefile is where caching occurs if possible.

\code{fetchData.file} currently does nothing. It exists to communicate that
the .file type is an option for data items.

\code{fetchData.sciencebase} downloads a file from ScienceBase.

\code{fetchData.url} downloads a file from the specified URL.
}

