% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{fetchTimestamp}
\alias{fetchTimestamp}
\alias{fetchTimestamp.sciencebase}
\title{Update timestamp file if needed}
\usage{
fetchTimestamp(viz.id, ...)

\method{fetchTimestamp}{sciencebase}(viz.id, data.info, old.timestamp,
  timestamp.file)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{data.info}{content information for this viz.id from the viz.yaml}
}
\description{
This function should be called from the generic, \code{fetchTimestamp()}.
Update timestamp file if the local copy of a file is out of date relative to 
a remote copy. The need for a timestamp file, and the method for finding the 
timestamp, will be determined by the metadata in viz.yaml for the item.

\code{fetchTimestamp.sciencebase} gets the file timestamp from ScienceBase.
}

