% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{fetchTimestamp}
\alias{fetchTimestamp}
\alias{fetchTimestamp.character}
\alias{fetchTimestamp.sciencebase}
\alias{fetchTimestamp.file}
\alias{fetchTimestamp.url}
\alias{fetchTimestamp.fetcher}
\title{Update timestamp file if needed}
\usage{
fetchTimestamp(viz)

\method{fetchTimestamp}{character}(viz)

\method{fetchTimestamp}{sciencebase}(viz)

\method{fetchTimestamp}{file}(viz)

\method{fetchTimestamp}{url}(viz)

\method{fetchTimestamp}{fetcher}(viz)
}
\arguments{
\item{viz}{the identifier for this data item in viz.yaml}
}
\value{
logical \code{TRUE} if a timestamp does not exist or has changed,
\code{FALSE} if the timestamps are identical.
}
\description{
This function should be called from the generic, \code{fetchTimestamp()}.
Update timestamp file if the local copy of a file is out of date relative to
a remote copy. The need for a timestamp file, and the method for finding the
timestamp, will be determined by the metadata in viz.yaml for the item.

\code{fetchTimestamp.character} fetches timestamp for a given id

\code{fetchTimestamp.sciencebase} gets the file timestamp from ScienceBase.

Does nothing - \code{make} already checks timestamps of local files

check a URL for timestamp

\code{fetchTimestamp.fetcher} superclass method catches missing
implementation
}
