% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{fetchTimestamp}
\alias{fetchTimestamp}
\alias{fetchTimestamp.character}
\alias{fetchTimestamp.fetcher}
\alias{fetchTimestamp.file}
\alias{fetchTimestamp.sciencebase}
\alias{fetchTimestamp.url}
\title{Update timestamp file if needed}
\usage{
fetchTimestamp(viz)

\method{fetchTimestamp}{character}(viz)

\method{fetchTimestamp}{sciencebase}(viz)

\method{fetchTimestamp}{file}(viz)

\method{fetchTimestamp}{url}(viz)

\method{fetchTimestamp}{fetcher}(viz)
}
\arguments{
\item{viz}{the identifier for this data item in viz.yaml}
}
\description{
This function should be called from the generic, \code{fetchTimestamp()}.
Update timestamp file if the local copy of a file is out of date relative to
a remote copy. The need for a timestamp file, and the method for finding the
timestamp, will be determined by the metadata in viz.yaml for the item.

\code{fetchTimestamp.sciencebase} gets the file timestamp from ScienceBase.

Has a file changed

check a URL for timestamp
}

