% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfileInfo.R
\name{findProfileYaml}
\alias{findProfileYaml}
\title{find where the profile.yaml file is saved.}
\usage{
findProfileYaml()
}
\description{
Only 2 directories supported at this point (home and relative).
}

