% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{generateWebpackConfigFile}
\alias{generateWebpackConfigFile}
\title{Create the webpack.config.js file
Needs an input and output in webpack.config}
\usage{
generateWebpackConfigFile(webpack.config)
}
\description{
Could switch out the template to a custom one.
}
