% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentInfo.R
\name{getBlocks}
\alias{getBlocks}
\title{Read vizlab config document and return specific blocks}
\usage{
getBlocks(block = c("vizlab", "info", "parameter", "fetch", "process",
  "visualize", "publish", "resource"), keep.block = TRUE)
}
\arguments{
\item{block}{character vector of blocks to read}

\item{keep.block}{logical whether to include block information in read block}
}
\description{
Read vizlab config document and return specific blocks
}
