% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfileInfo.R
\name{getProfileInfo}
\alias{getProfileInfo}
\title{Get information on profile settings}
\usage{
getProfileInfo()
}
\description{
Reads all values from the user's profile.yaml (path found through findProfileYaml)
}
\seealso{
findProfileYaml
}

