% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeBlockMakefile}
\alias{makeBlockMakefile}
\title{Create a single makefile from the information in viz.yaml}
\usage{
makeBlockMakefile(block = c("fetch", "process", "visualize", "publish"),
  outfile)
}
\arguments{
\item{block}{character name of the block in the viz.yaml for which to make a
makefile}

\item{outfile}{filename where the makefile should be written}
}
\description{
Uses information in the corresponding block of viz.yaml to create the
makefiles
}

