% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakefiles.R
\name{makeMakeItem}
\alias{makeMakeItem}
\title{Make a collection of makefile rules appropriate to a data/figure item}
\usage{
makeMakeItem(item.info, ...)
}
\arguments{
\item{item.info}{viz.yaml item info as from \code{getContentInfo}}
}
\description{
Make a collection of makefile rules appropriate to a data/figure item
}

