% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\alias{process.character}
\alias{process.unzip}
\title{Process raw or intermediate data products into intermediate or final data
products}
\usage{
process(viz)

\method{process}{character}(viz)

\method{process}{unzip}(viz.id, ..., outfile)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{...}{other arguments passed to process methods. These should
include any data dependencies, named according to the viz.id of those data
items}

\item{outfile}{the filename to which the processed data should be written}
}
\description{
This function should be called from the generic, \code{process()}.
Process raw or intermediate data products into intermediate or final data
products

\code{process.unzip} unzip a zip file
}

