% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\alias{publish.character}
\alias{publish.page}
\alias{publish.section}
\alias{publish.resource}
\alias{publish.img}
\alias{publish.ico}
\alias{publish.js}
\alias{publish.css}
\alias{publish.svg}
\alias{publish.footer}
\alias{publish.landing}
\alias{publish.template}
\title{Peform the publish step to ready the viz for hosting}
\usage{
publish(viz)

\method{publish}{character}(viz)

\method{publish}{page}(viz)

\method{publish}{section}(viz)

\method{publish}{resource}(viz)

\method{publish}{img}(viz)

\method{publish}{ico}(viz)

\method{publish}{js}(viz)

\method{publish}{css}(viz)

\method{publish}{svg}(viz)

\method{publish}{footer}(viz)

\method{publish}{landing}(viz)

\method{publish}{template}(viz)
}
\arguments{
\item{viz}{vizlab object or identifier}
}
\description{
Determine the type and dispatch to that method to produce
files to serve up as the final viz

publish a given id

publish a page

publish a section

publish a resource

Image publishing

Favicon resource

javascript publishing
TODO allow for cdn js

css publishing

svg publishing, may return NULL

Footer publishing

publish landing page

publish template
}
\details{
I've been thinking maybe switching 'x' to 'viz' would be more clear

Sections are published individually but are returned invisibly
as text to be used directly.

This copies static resources to the target directory, and invisibly
will return the preferred usage.

The job of minification or css precompiling could also be added here, but
currently this is not handled.

Also, templating the resources that make sense would be useful
}
