% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{publishLandingPage}
\alias{publishLandingPage}
\title{Get information about other visualizations from VIZLAB to generate the landing page}
\usage{
publishLandingPage(org = "USGS-VIZLAB", index_loc = "./output")
}
\arguments{
\item{org}{character, name of GitHub organization in which to pull out repository names}

\item{index_loc}{character, file path for where to save the index.html file}
}
\description{
note for future: github package is from GitHub repo rgithub (https://github.com/cscheid/rgithub)
there are plans to move the package to CRAN and call it grithub (https://github.com/cscheid/rgithub/issues/67)
}

