% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishLandingPage.R
\name{publishLandingPage}
\alias{publishLandingPage}
\title{Get information about other visualizations from VIZLAB to generate the landing page}
\usage{
publishLandingPage()
}
\description{
note for future: github package is from GitHub repo rgithub (https://github.com/cscheid/rgithub)
there are plans to move the package to CRAN and call it grithub (https://github.com/cscheid/rgithub/issues/67)
}

