% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\alias{readData.csv}
\alias{readData.default}
\alias{readData.excel}
\alias{readData.yaml}
\title{Ensure data is on the local file system}
\usage{
readData(viz.id, ...)

\method{readData}{default}(viz.id, data.info, ...)

\method{readData}{csv}(viz.id, data.info, ...)

\method{readData}{yaml}(viz.id, data.info, ...)

\method{readData}{excel}(viz.id, data.info, ...)
}
\arguments{
\item{viz.id}{the identifier for this data item in viz.yaml}

\item{...}{other arguments passed to readData methods}

\item{data.info}{content information for this viz.id from the viz.yaml}
}
\description{
This function should be called from the generic, \code{readData()}. Reads
data from a file into R format.

\code{readData.csv} reads a csv file.

\code{readData.yaml} reads a yaml file.

\code{readData.excel} reads the first spreadsheet of an Excel file.
}

