% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\alias{readData.character}
\alias{readData.excel}
\alias{readData.filepath}
\alias{readData.folder}
\alias{readData.rds}
\alias{readData.tabular}
\alias{readData.txt}
\alias{readData.yaml}
\title{Ensure data is on the local file system}
\usage{
readData(viz)

\method{readData}{character}(viz)

\method{readData}{tabular}(viz)

\method{readData}{yaml}(viz)

\method{readData}{excel}(viz)

\method{readData}{rds}(viz)

\method{readData}{filepath}(viz)

\method{readData}{folder}(viz)

\method{readData}{txt}(viz)
}
\arguments{
\item{viz}{vizlab object or identifier}
}
\description{
This function should be called from the generic, \code{readData()}. Reads
data from a file into R format. If you just want the filepath so you can do
it yourself, specify "customReader: filepath" in the viz.yaml

\code{readData.character} is the standard entry point; from here, viz.id gets
assigned a class to route it to a more specific reader

\code{readData.tabular} reads a text file.

\code{readData.yaml} reads a yaml file.

\code{readData.excel} reads the first spreadsheet of an Excel file.

\code{readData.RDS} reads an R object saved as an RDS.

\code{readData.filepath} returns the file path

\code{readData.folder} returns names of files inside a folder

\code{readData.txt} returns names of files inside a folder
}

