% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\alias{visualize.character}
\title{Visualize raw or intermediate data products into intermediate or final data
products}
\usage{
visualize(viz)

\method{visualize}{character}(viz)
}
\arguments{
\item{viz}{vizlab object described in viz.yaml}
}
\description{
This function should be called from the generic, \code{visualize()}.
Visualize raw or intermediate data products into intermediate or final data
products
}
