% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpacker.R
\name{webpacker}
\alias{webpacker}
\title{Run webpack for the project}
\usage{
webpacker(webpack.config)
}
\arguments{
\item{webpack.config}{list containing at least an input and output element that contain a
filepath for the javasript file going into webpack, and the name for the bundled file that results.}
}
\description{
This is called at the very end of \code{vizmake()} if you have a webpack block in your \code{viz.yaml}.
}
