% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTimestamp.R
\name{writeTimestamp}
\alias{writeTimestamp}
\title{Write a timestamp file with the conventions used by fetchTimestamp}
\usage{
writeTimestamp(new.timestamp, viz)
}
\arguments{
\item{new.timestamp}{the new timestamp to write to file}

\item{outfile}{the filename where the new timestamp should be saved}
}
\description{
Should only be used outside the vizlab package if the developer is writing a
new data + timestamp fetcher
}
